Add files via upload
TableSetup is the code that establishes a connection to a local database and creates a table called carlot and fills it with some car data.  feel free to change if needed

Car.java was my attempt at adding the car data by reading it from an external file and i can get the information read and shown in the document but the objects don't load as they should. If anyone wants to take a look at it go ahead.